+++ main.js
@@ -3,16 +3,26 @@ const app = express()
 const port = 3000
 
 const driver = require('./node/driver')
-const db = require('./db/db_insert')
+const dbIns = require('./db/db_insert')
+const dbSel = require('./db/db_select')
+const dbUpd = require('./db/db_update')
 
 
 
 app.use(express.static('/website/'))
 
-app.post('/driver/Login', (req,res) => res.send(driver.login("email","password")));
+app.post('/driver/Login', function (req,res) { 
+	var ret = dbSel.driverAuth(req.query.email, 
+						req.query.psw, 
+						function(results) { 
+						console.log(results);
+						res.send(results)});
+	 console.log("back to main");
+	 
+});
 app.get('/', (req, res) => req.send(index.html));
 app.post('/rest/Reg', function (req, res) {
-	var ret = db.restReg(req.query.uname, 
+	var ret = dbIns.restReg(req.query.uname, 
 						req.query.psw, 
 						req.query.name, 10, 
 						11, "(408)379-3333", 
@@ -22,7 +32,19 @@ app.post('/rest/Reg', function (req, res) {
 	 console.log("back to main");
 	 
 });
-													
+app.post('/driver/SignUp', function (req, res) {
+	var ret = dbIns.driverReg(req.query.uname, 
+						req.query.psw, 
+						req.query.name, 10, 
+						11, "(408)379-3333",
+						1.00, 
+						function(results) { 
+						console.log(results);
+						res.send(results)});
+	 console.log("back to main");
+	 
+});
+														
     
 
 app.listen(port, () => console.log(`Example app listening on port ${port}!`));
